refactor: consistent use of span as an exclusive range
i.e. span.hi is index of byte after lexeme.

span.hi used to be the index of the last byte of the lexeme (i.e.
inclusive range). I then decided to make it an exclusive range but
didn't fully commit to that, and some problems occurred wit EOF.
Finally, it was also hacky, and hard to figure out which convention was
used. This commit fixes this and also introduces a few abstractions to
make it easier to refactor/modify later on, if needed again.

Ideally, the output module wouldn't ever have to deal with spans, and
that logic would be entirely handled by Parser/Lexer so that it becomes
truly independent.